fix: fix overflow bug when `max_value = 255`.

+++ src/netpbm/pgm.rs
@@ -60,7 +60,7 @@ impl Serializable for PGM {
                 let grays = self
                     .data
                     .iter()
-                    .map(|x| x * (256 / (self.max_value + 1) as u16) as u8)
+                    .map(|x| x * (256 / (self.max_value as u16 + 1) as u16) as u8)
                     .collect::<Vec<_>>();
                 bytes.extend(grays);
             }
